requirements.txt
# React2Shell Hunter - Python Dependencies
# CVE-2025-55182 & CVE-2025-66478 Detection Toolkit

# AWS SDK
boto3>=1.34.0,<1.36.4  # Compatible with aiobotocore 2.19.0
botocore>=1.36.0,<1.36.4  # Compatible with aiobotocore 2.19.0
urllib3>=2.5.0  # Fixed: Open Redirect vulnerabilities

# Configuration
PyYAML>=6.0.1

# Data processing
pandas>=2.0.0
pyarrow>=14.0.0  # For Parquet/Athena integration

# Logging and monitoring
python-json-logger>=3.3.0  # Fixed: SNYK-PYTHON-PYTHONJSONLOGGER-9376713

# CLI interface
argparse>=1.4.0
click>=8.1.0

# Testing
pytest>=8.2,<9  # Compatible with pytest-asyncio 1.2.0
pytest-cov>=4.1.0
moto>=4.2.0  # AWS mocking

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
pylint>=3.1,<4  # Compatible with spyder 6.0.7
python-hcl2>=4.3.0  # For parsing Terraform files in tests

# Optional: Enhanced reporting
jinja2>=3.1.6  # Fixed: XSS & Template Injection vulnerabilities
tabulate>=0.9.0  # Table formatting
